refactor: Specify the extension of the file when importing non js/jsx
